moved services global variables to __init__.py. Updated code base where needed.
