+++ app/services/predictor.py
@@ -5,7 +5,7 @@ import pandas as pd
 import pytest
 
 from app.config import __model_version__
-from app.services.config_services import (
+from app.services import (
     BOOKING_MAP,
     MONTH_ABBREVIATION_MAP,
     VARIABLE_RENAME_MAP,
