+++ app/services/preprocessing.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Tuple
 import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
 
-from app.services.config_services import (
+from app.services import (
     BOOKING_MAP,
     MONTH_ABBREVIATION_MAP,
     VARIABLE_RENAME_MAP,
