+++ requirements.txt
@@ -1,13 +1,61 @@
-fastapi
-isort
-black
-pytest
-pytest-cov
-simplejson
-scikit-learn==1.5.2 
-feature-engine
-xgboost
-pytest-mock
-pytest-env
-requests
-httpx>=0.24.0
+annotated-types==0.7.0
+anyio==4.8.0
+asttokens==3.0.0
+black==24.10.0
+certifi==2024.12.14
+charset-normalizer==3.4.1
+click==8.1.8
+coverage==7.6.10
+decorator==5.1.1
+executing==2.1.0
+fastapi==0.115.6
+feature-engine==1.8.2
+h11==0.14.0
+httpcore==1.0.7
+httpx==0.28.1
+idna==3.10
+iniconfig==2.0.0
+ipython==8.31.0
+isort==5.13.2
+jedi==0.19.2
+joblib==1.4.2
+matplotlib-inline==0.1.7
+mypy-extensions==1.0.0
+numpy==2.2.1
+packaging==24.2
+pandas==2.2.3
+parso==0.8.4
+pathspec==0.12.1
+patsy==1.0.1
+pexpect==4.9.0
+platformdirs==4.3.6
+pluggy==1.5.0
+prompt_toolkit==3.0.48
+ptyprocess==0.7.0
+pure_eval==0.2.3
+pydantic==2.10.4
+pydantic_core==2.27.2
+Pygments==2.19.1
+pytest==8.3.4
+pytest-cov==6.0.0
+pytest-env==1.1.5
+pytest-mock==3.14.0
+python-dateutil==2.9.0.post0
+pytz==2024.2
+requests==2.32.3
+scikit-learn==1.5.2
+scipy==1.15.0
+simplejson==3.19.3
+six==1.17.0
+sniffio==1.3.1
+stack-data==0.6.3
+starlette==0.41.3
+statsmodels==0.14.4
+threadpoolctl==3.5.0
+traitlets==5.14.3
+typing_extensions==4.12.2
+tzdata==2024.2
+urllib3==2.3.0
+uvicorn==0.34.0
+wcwidth==0.2.13
+xgboost==2.1.3
