+++ tests/test_services/test_data_management.py
@@ -9,7 +9,7 @@ from feature_engine.encoding import OneHotEncoder
 from feature_engine.imputation import CategoricalImputer
 from sklearn.ensemble import RandomForestClassifier
 
-from app.services.config_services import VARS_TO_IMPUTE, VARS_TO_OHE
+from app.services import VARS_TO_IMPUTE, VARS_TO_OHE
 from app.services.pipeline import NoVacancyPipeline
 from app.services.preprocessing import NoVacancyDataProcessing
